Added blender versions of chess pieces and added one to the scene
The position and rotation of the pieces is weird. I’m not sure  if this
is blender or the original settings of the 3D object. Objects also have
a cube in the model that we can maybe use to reposition them in Blender
but I’m not familiar with how to do that.